Avoided the repetitive registering and deregistering event source as suggested by TLama
I would prefer their methods to be the instance methods rather than
class methods to avoid the repetitive registering and deregistering
event source. I would register the event source when the instance of the
class is created and deregister when destroyed. Or make a global
threadvar and initialize it once.

http://stackoverflow.com/questions/30229826/writing-to-the-windows-event-log-using-delphi/30229827?noredirect=1#comment48566147_30229827